user 기능 추가

+++ articles/views.py
@@ -23,7 +23,9 @@ def create(request):
             # image = form.cleaned_data.get("image")
             # article = Article(title=title, content=content, image=image)
             # article.save()
-            article = form.save()
+            article = form.save(commit=False)
+            article.user_id = request.user.id
+            article.save()
             return redirect("articles:detail", article.pk)
     else:
         form = ArticleForm()
@@ -50,7 +52,8 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if article.user == request.user:
+        article.delete()
     return redirect("articles:index")
     # if request.method == "POST":
     #     article.delete()
@@ -61,21 +64,24 @@ def delete(request, article_pk):
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST, request.FILES, instance=article)
-        if form.is_valid():
-            # article.title = form.cleaned_data.get("title")
-            # article.content = form.cleaned_data.get("content")
-            # if form.cleaned_data.get("image"):
-            #     article.image = form.cleaned_data.get("image")
-            # article.save()
-            form.save()
-            return redirect("articles:detail", article.pk)
+    if article.user == request.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST, request.FILES, instance=article)
+            if form.is_valid():
+                # article.title = form.cleaned_data.get("title")
+                # article.content = form.cleaned_data.get("content")
+                # if form.cleaned_data.get("image"):
+                #     article.image = form.cleaned_data.get("image")
+                # article.save()
+                form.save()
+                return redirect("articles:detail", article.pk)
+        else:
+            # form = ArticleForm(initial=article.__dict__)
+            form = ArticleForm(instance=article)
+        context = {"form": form, "article": article}
+        return render(request, "articles/form.html", context)
     else:
-        # form = ArticleForm(initial=article.__dict__)
-        form = ArticleForm(instance=article)
-    context = {"form": form, "article": article}
-    return render(request, "articles/form.html", context)
+        return redirect("articles:index")
 
 
 @require_POST
@@ -87,6 +93,7 @@ def comments_create(request, article_pk):
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article_id = article_pk
+        comment.user_id = request.user.id
         comment.save()
     return redirect("articles:detail", article_pk)
     # if request.method == "POST":
@@ -105,7 +112,8 @@ def comments_create(request, article_pk):
 @login_required
 def comments_delete(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if comment.user == request.user:
+        comment.delete()
     return redirect("articles:detail", article_pk)
     # if request.method == "POST":
     #     comment = get_object_or_404(Comment, pk=comment_pk)
@@ -116,16 +124,19 @@ def comments_delete(request, article_pk, comment_pk):
 @login_required
 def comments_update(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    if request.method == "POST":
-        # comment.content = request.POST.get("content")
-        # comment.save()
-        comment_form = CommentForm(request.POST, instance=comment)
-        if comment_form.is_valid():
-            comment = comment_form.save(commit=False)
-            comment.article_id = article_pk
-            comment.save()
-            return redirect("articles:detail", article_pk)
+    if comment.user == request.user:
+        if request.method == "POST":
+            # comment.content = request.POST.get("content")
+            # comment.save()
+            comment_form = CommentForm(request.POST, instance=comment)
+            if comment_form.is_valid():
+                comment = comment_form.save(commit=False)
+                comment.article_id = article_pk
+                comment.save()
+                return redirect("articles:detail", article_pk)
+        else:
+            comment_form = CommentForm(instance=comment)
+        context = {"comment_form": comment_form}
+        return render(request, "articles/comment_update.html", context)
     else:
-        comment_form = CommentForm(instance=comment)
-    context = {"comment_form": comment_form}
-    return render(request, "articles/comment_update.html", context)
+        return redirect("articles:detail", article_pk)
